REFACTOR: upload policy class
