+++ sample/src/main/java/com/imagekit/android/sample/UploadImageActivity.kt
@@ -3,10 +3,8 @@ package com.imagekit.android.sample
 import android.app.Activity
 import android.app.AlertDialog
 import android.content.Intent
-import android.content.IntentFilter
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
-import android.os.BatteryManager
 import android.os.Bundle
 import android.util.Log
 import android.view.View
@@ -14,7 +12,7 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.imagekit.android.ImageKit
 import com.imagekit.android.ImageKitCallback
-import com.imagekit.android.UploadPolicy
+import com.imagekit.android.entity.UploadPolicy
 import com.imagekit.android.entity.UploadError
 import com.imagekit.android.entity.UploadResponse
 import kotlinx.android.synthetic.main.activity_upload_image.*
